api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
openai==1.12.0
python-dotenv==1.0.0
python-multipart==0.0.7
langchain==0.1.1
langchain-openai==0.0.5
python-jose==3.3.0
passlib==1.7.4
